I added so many things without ever committing them that now I honestly have no idea what I did except for maybe the last thing or two. Added the map feature for viewing events, also added the popup marker on the map. Also started implementing the bottomSheet for search parameters input.
Yes this is not the right branch for all of this, I'll take care of it tomorrow.

+++ lib/screens/booked_events_screen.dart
@@ -142,7 +142,7 @@ class _BookedEventsScreenState extends State<BookedEventsScreen> {
                 return EventItem(
                   suggested[index],
                   index,
-                  dummy.length.toDouble(),
+                  dummy.length,
                 );
               },
               childCount:
